+++ src/test/java/de/sephirothj/spring/security/ltpa2/Ltpa2ConfigurerTest.java
@@ -42,8 +42,10 @@ class Ltpa2ConfigurerTest
 	@Test
 	void testConfigure() throws Exception
 	{
+		ApplicationContext context = mock(ApplicationContext.class);
 		HttpSecurity httpSecurity = mock(HttpSecurity.class);
-		given(httpSecurity.getSharedObject(UserDetailsService.class)).will(invocation -> mock(invocation.getArgument(0)));
+		given(httpSecurity.getSharedObject(ApplicationContext.class)).will(invocation -> context);
+		given(context.getBean(UserDetailsService.class)).will(invocation -> mock(invocation.getArgument(0)));
 		final String headerName = "header";
 		final String cookieName = "cookie";
 		final SecretKey sharedKey = LtpaKeyUtils.decryptSharedKey(Constants.ENCRYPTED_SHARED_KEY, Constants.ENCRYPTION_PASSWORD);
@@ -73,8 +75,10 @@ class Ltpa2ConfigurerTest
 	@Test
 	void testConfigureWithAuthFaulureHandler() throws Exception
 	{
+		ApplicationContext context = mock(ApplicationContext.class);
 		HttpSecurity httpSecurity = mock(HttpSecurity.class);
-		given(httpSecurity.getSharedObject(UserDetailsService.class)).will(invocation -> mock(invocation.getArgument(0)));
+		given(httpSecurity.getSharedObject(ApplicationContext.class)).will(invocation -> context);
+		given(context.getBean(UserDetailsService.class)).will(invocation -> mock(invocation.getArgument(0)));
 		final SecretKey sharedKey = LtpaKeyUtils.decryptSharedKey(Constants.ENCRYPTED_SHARED_KEY, Constants.ENCRYPTION_PASSWORD);
 		final PublicKey publicKey = LtpaKeyUtils.decodePublicKey(Constants.ENCODED_PUBLIC_KEY);
 		final AuthenticationFailureHandler failureHandler = new AuthenticationEntryPointFailureHandler(new Http403ForbiddenEntryPoint());